Fix Godot error caused by the terrain plugin handling more than one type.
This was causing multiple plugins to handle the same object, which Godot
is not expecting by design.
(https://github.com/godotengine/godot/issues/73650#issuecomment-1811506274)
It is still quite a hole in the API, because nested editing contextes are
happening very often in this project, and need to be handled somehow.
Errors are also really obscure as they lack the details that would
help figuring out these problems.

VoxelTerrainEditorPlugin was handling all these types just to avoid
edit(null) and make_visible(false) which would make its tools disappear
when selecting anything else, which is not desired when the object is
actually part of the terrain's configuration (modifiers, generators...).
Turns out, removing this kept behavior decent enough. For some reason
inspecting the generator didn't make terrain controls disappear.
However, they might disappear with modifiers now, but we'll see later if
we need to bring back some workaround for them.