Fix Godot error caused by the graph plugin handling more than one type.
Similar to previous commit fixing similar errors. This case has the same
root cause: preventing tools from getting turned off when inspecting
another related object, but because it was attempted by having the plugin
handle those objects too in the `handles` method, it somehow confused
Godot.

However this needed a different fix, as we're not dealing with scene nodes.
The error happened when inspecting the graph editor after having
inspected one of its nodes. It's unclear why exactly that order was
problematic and not the opposite, but removing nodes from handled types
fixed it. However it meant that the graph editor would completely close
everytime a node of the graph is selected. Since node inspection is
triggered by this plugin, we can workaround that by ignoring edit(null)
AND make_visible(false) around the call to inspect_object.
However this still breaks if the navigation arrows of the inspector
are used. It's also unclear why `inspector_only=true` doesn't help in this
case.